수정 - socket 설정 모듈화
