+++ package.json
@@ -26,7 +26,8 @@
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "socket.io": "^4.5.1",
-    "ts-node": "^10.8.1"
+    "ts-node": "^10.8.1",
+    "unique-names-generator": "^4.7.1"
   },
   "devDependencies": {
     "nodemon": "^2.0.16"
