+++ src/server.ts
@@ -3,6 +3,7 @@ import cors from 'cors';
 import 'dotenv/config';
 import { Server } from 'socket.io';
 import { chatRouter } from './routes';
+import webSocket from './static/ts';
 
 const app = express();
 app.use(cors());
@@ -21,27 +22,4 @@ app.use('/chat', chatRouter);
 const server = app.listen(PORT, () => console.log(`server is running ${PORT}`));
 
 // // socket
-const io = new Server(server);
-
-io.sockets.on('connection', (socket) => {
-  console.log('socket 접속');
-
-  socket.on('newUser', (newUser) => {
-    socket.emit('update', `${newUser}(나)님이 접속했습니다.`);
-    socket.broadcast.emit('update', `${newUser}님이 접속했습니다.`);
-    (socket as any).name = newUser;
-  });
-
-  socket.on('chat message', (msg: any) => {
-    const name = (socket as any).name;
-    console.log(msg);
-    // 다른사람 화면에 보여지는 메세지
-    socket.broadcast.emit('chat message', `${msg.name}: ${msg.msg}`);
-    // 내 화면에 보여지는 메세지
-    socket.emit('chat message', `나: ${msg.msg}`);
-  });
-
-  socket.on('disconnect', () => {
-    socket.broadcast.emit('update', `${(socket as any).name}님이 나가셨습니다.`);
-  });
-});
+webSocket(server);
