+++ src/static/index.html
@@ -62,7 +62,6 @@
     <script>
       var socket = io();
       var name = prompt('반갑습니다!', '');
-
       if (!name) {
         name = '익명';
       }
